+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - Improved Status logging. [#742](https://github.com/pulumi/pulumi-kubernetes-operator/pull/742)
 - Support for ReconcileRequest annotation. [#745](https://github.com/pulumi/pulumi-kubernetes-operator/pull/745)
 - Show stack processing state in printer columns. [#747](https://github.com/pulumi/pulumi-kubernetes-operator/pull/747)
+- Some stack outputs cannot be saved to Secret. [#746](https://github.com/pulumi/pulumi-kubernetes-operator/pull/746)
 
 ## 2.0.0-beta.1 (2024-10-18)